feat: fit reduce duplication
